refactor: CustomizePage 여러 컴포넌트로 쪼개기
